refactor(editor): Reorganize shims-xxx.d.ts files (no-changelog)
