+++ packages/frontend/editor-ui/src/app/shims-global.d.ts
@@ -1,4 +1,12 @@
+/// <reference types="vite/client" />
+/// <reference types="vite-plugin-comlink/client" />
+
 import 'vue-router';
+import type { VNode, ComponentPublicInstance } from 'vue';
+import type { PartialDeep } from 'type-fest';
+import type { ExternalHooks } from '@/app/types/externalHooks';
+import type { FrontendSettings } from '@n8n/api-types';
+import type { Plugin as PrettierPlugin } from 'prettier';
 import type { I18nClass } from '@n8n/i18n';
 import type { Route, Router, RouteLocation } from 'vue-router';
 import type { Telemetry } from '@/app/plugins/telemetry';
@@ -6,9 +14,63 @@ import type { VIEWS } from '@/app/constants';
 import type { IPermissions } from '@/Interface';
 import type { MiddlewareOptions, RouterMiddlewareType } from '@/app/types/router';
 
+/**
+ * File types
+ */
+
+declare module '*.json';
+declare module '*.svg';
+declare module '*.png';
+declare module '*.jpg';
+declare module '*.jpeg';
+declare module '*.gif';
+declare module '*.webp';
+
+declare module '*?raw' {
+	const content: string;
+	export default content;
+}
+
+/**
+ * Global
+ */
+
+declare global {
+	interface ImportMeta {
+		env: {
+			DEV: boolean;
+			PROD: boolean;
+			NODE_ENV: 'development' | 'production';
+			VUE_APP_URL_BASE_API: string;
+			VUE_SCAN: boolean;
+		};
+	}
+
+	interface Window {
+		BASE_PATH: string;
+		REST_ENDPOINT: string;
+		n8nExternalHooks?: PartialDeep<ExternalHooks>;
+		preventNodeViewBeforeUnload?: boolean;
+		maxPinnedDataSize?: number;
+	}
+
+	namespace JSX {
+		interface Element extends VNode {}
+		interface ElementClass extends ComponentPublicInstance {}
+		interface IntrinsicElements {
+			[elem: string]: any;
+		}
+	}
+
+	interface Array<T> {
+		findLast(predicate: (value: T, index: number, obj: T[]) => unknown, thisArg?: any): T;
+	}
+}
+
 export {};
 
 /**
+ * Vue Runtime
  * @docs https://vuejs.org/guide/typescript/options-api.html#augmenting-global-properties
  */
 
@@ -28,6 +90,7 @@ declare module '@vue/runtime-core' {
 }
 
 /**
+ * Vue Router
  * @docs https://router.vuejs.org/guide/advanced/meta
  */
