--- packages/frontend/editor-ui/src/vite-virtual-nodes-popularity.d.ts
@@ -1,9 +1,0 @@
-/// <reference types="vite/client" />
-
-declare module 'virtual:node-popularity-data' {
-	const data: Array<{
-		id: string;
-		popularity: number;
-	}>;
-	export default data;
-}
